Clean up more backend fetches
