--- yky-social/yky/app/components/SendTimeline.tsx
@@ -1,60 +1,0 @@
-//import React, { useEffect, useState } from 'react';
-import React from 'react';
-import { getuserid } from './userid';
-import { getAPIServer } from './backend';
-
-//{postId: tle.post_id,  fromUserId:tle.user_id, sendDate: tle.send_date, sendType:tle.send_type,
-//    postText: tle.post?.text, postMentions: tle.post?.mentions}
- 
-interface SendItem {
-    postId: string;
-    postText: string;
-}
-
-interface Props {
-  userid: string;
-}
-
-const SendTimeline: React.FC<Props> = async ({ userid }) => {
-    const luserid = getuserid();
-    let rquserid = userid;
-    if (!rquserid || rquserid === 'default') {
-      rquserid = luserid;
-    }
-  
-    try {
-        const response = await fetch(getAPIServer() + '/sendtimeline'+'?' + new URLSearchParams({
-            userid: luserid,
-            rqtimeline: rquserid,
-          }),
-        {
-          method: 'GET',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        });
-    
-        const data = await response.json();
-
-        const messageList = data.timeline as SendItem[];
-
-        return (
-          <div className="container mx-auto px-4 py-8">
-            <h1 className="text-2xl font-bold mb-4">Sent Messages and Posts</h1>
-            <ul className="space-y-4">
-              {messageList.map((message) => (
-                <li key={message.postId} className="bg-white p-4 rounded shadow">
-                  <p>{message.postText}</p>
-                </li>
-              ))}
-            </ul>
-          </div>
-        );
-    }
-    catch (error) {
-        console.error('Error fetching messages:', error);
-        return (<h1>Error!</h1>);
-    }
-};
-
-export default SendTimeline;
